make the docker matchers less IO hungry
this was achieved by making the following changes:

1. reordering the dockerfile so less common changes happen first
2. don't remove the intermediate containers and use the cache

this should solve some of the problems we are seeing on CI regarding
slow builds and IO hungry docker daemon. keep in mind that this could
lead to higher disk utilization which means we will have to run `docker
images | xargs docker rmi` frequently to delete old images.